+++ engine/src/hopeit/dataobjects/jsonify.py
@@ -12,6 +12,6 @@ class Json(Payload):
     Json convenience ser/deser functions for @dataobject decorated object (@see DataObject)
     """
     warnings.warn(
-        "Usages of `Json` from `hopeit.engine.jsonify` is deprecated since 0.8.3 and will be "
-        "removed in version 0.9.0,use `Payload` from `hopeit.engine.payload` instead", DeprecationWarning
+        "Usage of `Json` from `hopeit.engine.jsonify` is deprecated since 0.8.3 and will be removed "
+        "in a future version. Use `Payload` from `hopeit.engine.payload` instead", DeprecationWarning
     )